Actor movement is almost done (not fully tested)

+++ PHEngine/PHEngine/src/Core/GameCore/Components/ComponentData/DirectionalLightComponentData.h
@@ -19,7 +19,7 @@ namespace Game
       {
       }
 
-      virtual size_t GetType() override {
+      virtual uint64_t GetType() override {
 
          return DIR_LIGHT_COMPONENT;
       }
