+++ PHEngine/PHEngine/src/Core/GameCore/Components/ComponentData/SkyboxComponentData.h
@@ -23,7 +23,7 @@ namespace Game
 
 		}
 
-		virtual size_t GetType() override {
+		virtual uint64_t GetType() override {
 
 			return SKYBOX_COMPONENT;
 		}
