+++ PHEngine/PHEngine/src/Core/GameCore/Components/ComponentData/StaticMeshComponentData.h
@@ -26,7 +26,7 @@ namespace Game
 		{
 		}
 
-		virtual size_t GetType() override {
+		virtual uint64_t GetType() override {
 
 			return STATIC_MESH_COMPONENT;
 		}
